Particles: consistently use floats for particle data calculations.
This removes float<->double conversions, which makes the auto-vectorized
code even smaller.

+++ src/particles/qquickage.cpp
@@ -89,15 +89,15 @@ bool QQuickAgeAffector::affectParticle(QQuickParticleData *d, qreal dt)
 {
     Q_UNUSED(dt);
     if (d->stillAlive(m_system)){
-        qreal curT = (qreal)m_system->timeInt/1000.0;
-        qreal ttl = (qreal)m_lifeLeft/1000.0;
+        float curT = m_system->timeInt / 1000.0f;
+        float ttl = m_lifeLeft / 1000.0f;
         if (!m_advancePosition && ttl > 0){
-            qreal x = d->curX(m_system);
-            qreal vx = d->curVX(m_system);
-            qreal ax = d->curAX();
-            qreal y = d->curY(m_system);
-            qreal vy = d->curVY(m_system);
-            qreal ay = d->curAY();
+            float x = d->curX(m_system);
+            float vx = d->curVX(m_system);
+            float ax = d->curAX();
+            float y = d->curY(m_system);
+            float vy = d->curVY(m_system);
+            float ay = d->curAY();
             d->t = curT - (d->lifeSpan - ttl);
             d->setInstantaneousX(x, m_system);
             d->setInstantaneousVX(vx, m_system);
